+++ parser/sql/statement/type/hive/src/main/java/org/apache/shardingsphere/sql/parser/statement/hive/tcl/HiveAbortStatement.java
@@ -15,43 +15,17 @@
  * limitations under the License.
  */
 
-grammar HiveStatement;
+package org.apache.shardingsphere.sql.parser.statement.hive.tcl;
 
-import Comments, DMLStatement, DDLStatement, DALStatement, TCLStatement;
+import org.apache.shardingsphere.infra.database.core.type.DatabaseType;
+import org.apache.shardingsphere.sql.parser.statement.core.statement.type.dal.DALStatement;
 
-// TODO correct hive SQL parsing according to official documentation
-execute
-    : (select
-    | insert
-    | update
-    | delete
-    | loadStatement
-    | createDatabase
-    | dropDatabase
-    | alterDatabase
-    | use
-    | createTable
-    | dropTable
-    | truncateTable
-    | msckStatement
-    | alterTable
-    | createView
-    | dropView
-    | alterView
-    | createMaterializedView
-    | dropMaterializedView
-    | alterMaterializedView
-    | createIndex
-    | dropIndex
-    | alterIndex
-    | createMacro
-    | dropMacro
-    | createFunction
-    | dropFunction
-    | reloadFunction
-    | show
-    | describe
-    | abort
-    ) (SEMI_ EOF? | EOF)
-    | EOF
-    ;
+/**
+ * Hive abort statement.
+ */
+public final class HiveAbortStatement extends DALStatement {
+    
+    public HiveAbortStatement(final DatabaseType databaseType) {
+        super(databaseType);
+    }
+}